--- json_serializable/tool/watch.dart
@@ -1,12 +1,0 @@
-#!/usr/bin/env dart --checked
-// Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-import 'package:build_runner/build_runner.dart';
-
-import 'build_actions.dart';
-
-main() {
-  watch(builders, deleteFilesByDefault: true);
-}
